Undid some moving
